refactor: TravelRepository
